Refactor into modular package and add tests
