+++ README.md
@@ -1,40 +1,62 @@
 # DevAI-R1
-Sistema de inteligência de desenvolvimento com IAs.
 
-Esta atualização permite explorar todo o limite de 160k tokens
-oferecido pela API do OpenRouter para entrada e saída de dados.
-Agora o aplicativo define `MAX_CONTEXT_LENGTH` em `160000` e envia
-prompts para a API utilizando esse limite, garantindo respostas mais
-longas quando necessário.
+Assistente de desenvolvimento baseado em IA com suporte a contextos de até **160k tokens** via OpenRouter.
 
-## Novidades
+## Principais recursos
 
-- **Persistência de índice FAISS**: o índice de busca semântica é salvo em disco,
-  acelerando a inicialização.
-- **Limpeza automática de memórias**: entradas antigas ou pouco acessadas são
-  removidas diariamente para manter o banco enxuto.
-- **Tempo limite nas chamadas à API**: requisições ao OpenRouter agora possuem
-  timeout e contagem de erros para maior estabilidade.
-- **Métricas de uso**: número de chamadas à API e tempo médio de resposta estão
-  disponíveis em `/metrics`.
+- **Memória persistente** com busca vetorial (FAISS)
+- **Limpeza automática de memórias** e feedback de uso
+- **Análise de código** com construção de grafo de dependências
+- **Monitoramento de logs** detectando padrões de erro
+- **API FastAPI** e **interface de linha de comando**
+- Métricas expostas em `/metrics`
 
 ## Configuração
 
-Crie um arquivo `config.yaml` na raiz para personalizar caminhos e outras
-configurações. Exemplo:
+1. Crie um arquivo `config.yaml` na raiz (opcional) e defina os caminhos desejados:
 
 ```yaml
 CODE_ROOT: ./app
 API_PORT: 8000
 ```
 
-Variáveis de ambiente ainda podem sobrescrever `OPENROUTER_API_KEY`.
+2. Defina a variável de ambiente `OPENROUTER_API_KEY` com sua chave de acesso.
 
-## Monitoramento da pasta `./app`
+## Executando
 
-O aplicativo verifica periodicamente a pasta especificada em `CODE_ROOT` e se
-atualiza automaticamente sempre que novos arquivos Python são adicionados ou
-modificados.
+- **Servidor API**:
 
-Uma interface de administração em `/admin` permite acionar uma nova varredura
-manual e exibe informações sobre o último escaneamento.
+```bash
+python -m devai --api
+```
+
+- **Interface de linha de comando**:
+
+```bash
+python -m devai --cli
+```
+
+Os comandos disponíveis na CLI são listados ao iniciar o programa, como `/memoria`, `/tarefa` e `/grafo`.
+
+## Testes
+
+Instale as dependências de desenvolvimento e execute:
+
+```bash
+pytest
+```
+
+## Estrutura modular
+
+O código foi dividido em módulos dentro do pacote `devai/`, facilitando a inclusão de novas funcionalidades:
+
+- `config.py` – carregamento de configuração e logging
+- `memory.py` – gerenciamento de memórias
+- `analyzer.py` – análise de código e grafo de dependências
+- `tasks.py` – execução de tarefas
+- `log_monitor.py` – monitoramento de arquivos de log
+- `ai_model.py` – comunicação com o OpenRouter
+- `core.py` – orquestração principal
+- `cli.py` – interface de linha de comando
+
+Sinta‑se livre para expandir cada módulo conforme necessário.
